try block:
This block of code contains code that may cause an error. If an error occurs in the middle of a try block, code execution moves to a catch block.
const results = await getGeocode({ address: address.label }):

getGeocode is a function from the use-places-autocomplete library that converts an address into a geocode. await waits for getGeocode to complete and returns the result.
address.label is the text of the selected address. The result of getGeocode is an array of objects that contain geocoded address information.
catch (error) block:
If an error occurs in the try block, the try block is executed.