The options object sets various parameters for the Google Maps map. Added options:
disableDefaultUI: Disabling the default Google Maps interface.
draggable: Permission to move the map with the mouse.
gestureHandling: Configure gesture handling for the map so that any part of the map can move the map.
mapTypeControl: Enable a map type switch to be able to select the map type.
minZoom and maxZoom: The minimum and maximum zoom levels are set, which limit the zooming of the map to certain limits.

restriction: Restrictions are set for moving the map to the specified latitude and longitude limits. The strictBounds: true parameter specifies that the map cannot be moved beyond the set limits.